# perf: Optimize available-drivers-info