Since the idea to cache available-drivers-info (#55596) may not work, here's the attempt to optimize it. The main focus is on reducing allocations, getting rid of lazy sequences in favor of vectors, and being a bit smarter about the overall algorithm.

```
BEFORE:
Time per call: 1.27 ms   Alloc per call: 2,932,885b
Time per call: 1.23 ms   Alloc per call: 2,932,713b
Time per call: 1.23 ms   Alloc per call: 2,932,721b

AFTER:
Time per call: 463.75 us   Alloc per call: 888,376b
Time per call: 456.28 us   Alloc per call: 888,032b
Time per call: 458.39 us   Alloc per call: 888,032b
```

Allocation diffgraph: https://flamebin.dev/PTWcnh. Not terribly useful as it's just all blue:).